% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liz_coords}
\alias{liz_coords}
\title{Example coordinates}
\format{
A data frame with 2 columns
\describe{
\item{x}{Longitude coordinate.}
\item{y}{Latitude coordinate.}
}
}
\source{
Data from Bouzid et al. 2022 (DOI: 10.1111/mec.15836).
}
\usage{
liz_coords
}
\description{
Example coordinates
}
\keyword{datasets}
