% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liz_vcf}
\alias{liz_vcf}
\title{Example VCF}
\format{
Object of class vcfR with 101 individuals and 6,944 loci.
}
\source{
Data from Bouzid et al. 2022 (DOI: 10.1111/mec.15836).
}
\usage{
liz_vcf
}
\description{
A Variant Call Format data set
}
\keyword{datasets}
