% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbowpca.R
\name{pca_to_rgb}
\alias{pca_to_rgb}
\title{Convert three PCA axes to RGB colors}
\usage{
pca_to_rgb(x, order = c(1, 2, 3))
}
\arguments{
\item{x}{A numeric matrix or data frame}

\item{order}{Integer vector of length 3 specifying which columns should map
to the red, green, and blue channels, respectively.}
}
\value{
A tibble containing:
\describe{
\item{R}{Rescaled red channel}
\item{G}{Rescaled green channel}
\item{B}{Rescaled blue channel}
\item{color}{RGB color string}
}
}
\description{
Converts a numeric matrix (or data frame) into RGB colors
by independently rescaling each column to the range \link{0, 1}.
Defaults to plotting first three columns as red, green, and blue,
but the \code{order} argument allows you to specify which columns should
map to which channels.
}
\examples{
pcs <- matrix(rnorm(300), ncol = 3)

rgb_df <- pca_to_rgb(pcs)

# Swap channels
rgb_df2 <- pca_to_rgb(pcs, order = c(3, 1, 2))

}
