% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbowpca.R
\name{rescale_channel}
\alias{rescale_channel}
\title{Rescale numeric values to an RGB channel}
\usage{
rescale_channel(x, to = c(0, 1))
}
\arguments{
\item{x}{Numeric vector.}

\item{to}{Numeric vector of length 2. Output range.}
}
\value{
Numeric vector rescaled to \code{to}.
}
\description{
Rescale numeric values to an RGB channel
}
